change aspect ratio of 3d plot + bug fix in prop detection
In the old version of the code, it determined whether all the electrode activity fell into the same time bin by comparing the maximum activity value to the number of electrodes which became active. However, if only two electrodes contained activity, and they were both active in the same time bin, but also one (or more) of them was active in a separate time bin, it would not consider it as a propagation detection. This was fixed by considering whether the total activity for the entirety of the action potential is equal to the maximum activity for the entirety of the action potential.